ENH: support Cryptography for encryption/decryption
