+++ pyproject.toml
@@ -191,6 +191,7 @@ ignore = [
 "tests/*" = ["S101", "ANN001", "ANN201","D104", "S105", "S106", "D103", "B018", "B017"]
 "sample-files/*" = ["D100", "INP001"]
 "_encryption.py" = ["S324", "S311"]
+"_cryptography.py" = ["S304", "S305"]  # Use of insecure cipher / modes, aka RC4 and AES-ECB
 "_writer.py" = ["S324"]
 "make_changelog.py" = ["T201", "S603", "S607"]
 "json_consistency.py" = ["T201"]
